# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "image repository",
    "version": "1.0.210104",
    "title": "Image Repository",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "bruce.he.62@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "imgrepo.io",
  "basePath": "/",
  "schemes": ["https", "http"],
  "paths": {
    "/signup": {
      "post": {
        "summary": "sign up",
        "tags": ["API"],
        "description": "",
        "operationId": "signup",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "user credential",
            "description": "user's email and password",
            "in": "body",
            "schema": {
              "$ref": "#definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "bad request"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },
    "/images?type={type}": {
      "post": {
        "summary": "uploads an image",
        "tags": ["API"],
        "description": "",
        "operationId": "uploadImage",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "email",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "required": true,
            "type": "string",
            "format": "password"
          },
          {
            "name": "images",
            "in": "formData",
            "description": "file to upload",
            "required": true,
            "type": "file"
          },
          {
            "name": "type",
            "in": "path",
            "description": "image type, public or private, default is private",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "bad request"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },
    "/images": {
      "get": {
        "summary": "get image list",
        "tags": ["API"],
        "description": "",
        "operationId": "imageList",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "email",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "required": true,
            "type": "string",
            "format": "password"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "bad request"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },
    "/images/{id}?type={type}": {
      "get": {
        "summary": "download an image",
        "tags": ["API"],
        "description": "",
        "operationId": "downloadImage",
        "produces": ["image/jpeg", "image/png"],
        "parameters": [
          {
            "name": "email",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "required": false,
            "type": "string",
            "format": "password"
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of the image to return",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "path",
            "description": "returned image type, original or thumbnail ",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "bad request"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },
    "/images/{id}": {
      "delete": {
        "summary": "delete an image",
        "tags": ["API"],
        "description": "",
        "operationId": "deleteImage",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "email",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "required": true,
            "type": "string",
            "format": "password"
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of the image to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "bad request"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },

    "/images/search?q={q}": {
      "get": {
        "summary": "search images",
        "tags": ["API"],
        "description": "",
        "operationId": "searchImage",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "email",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "header",
            "required": true,
            "type": "string",
            "format": "password"
          },
          {
            "name": "q",
            "in": "path",
            "description": "search query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "bad request"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "required": ["email", "password"],
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "description": "User's email",
          "example": "bruce.he.62@gmail.com"
        },
        "password": {
          "type": "string",
          "format": "password",
          "description": "User's password",
          "example": "mypassword"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}